JetBrains annotations, further reorganization.
