+++ Core/Utilities/TypeConversionExtensions.cs
@@ -2,7 +2,7 @@
 
 namespace Zenith.Core.Utilities;
 
-public static class TypeConversionExtensions
+internal static class TypeConversionExtensions
 {
     public static Vector3 ToVector3(this Vector2 vector2) => new(vector2.X, vector2.Y, 0);
 }
