+++ ZenithMod.cs
@@ -1,9 +1,11 @@
+using JetBrains.Annotations;
 using Terraria.ModLoader;
 using Zenith.Core.Features.PrimitiveRendering;
 
 namespace Zenith;
 
-public class ZenithMod : Mod
+[UsedImplicitly(ImplicitUseKindFlags.InstantiatedWithFixedConstructorSignature)]
+public sealed class ZenithMod : Mod
 {
     public override void PostSetupContent()
     {
